RFT: Remove Position management from PerformancePeriod. This cuts down on the number of per-tick update that occur since they were duplicated per each PerformancePeriod. Also opens up the path to cythonizing the entire object
