+++ tests/test_perf_tracking.py
@@ -879,12 +879,15 @@ class TestPositionPerformance(unittest.TestCase):
 
         txn1 = create_txn(trades_1[1], 10.0, 100)
         txn2 = create_txn(trades_2[1], 10.0, -100)
-        pp = perf.PerformancePeriod(1000.0)
-        pp.execute_transaction(txn1)
-        pp.execute_transaction(txn2)
+        pt = perf.PositionTracker()
+        pp = perf.PerformancePeriod(1000.0, position_tracker=pt)
+        pt.execute_transaction(txn1)
+        pp.handle_execution(txn1)
+        pt.execute_transaction(txn2)
+        pp.handle_execution(txn2)
 
         for trade in itertools.chain(trades_1[:-2], trades_2[:-2]):
-            pp.update_last_sale(trade)
+            pt.update_last_sale(trade)
 
         pp.calculate_performance()
 
@@ -911,9 +914,9 @@ class TestPositionPerformance(unittest.TestCase):
                       net_liquidation=1000.0)
 
         # now simulate stock1 going to $9
-        pp.update_last_sale(trades_1[-1])
+        pt.update_last_sale(trades_1[-1])
         # and stock2 going to $11
-        pp.update_last_sale(trades_2[-1])
+        pt.update_last_sale(trades_2[-1])
 
         pp.calculate_performance()
 
@@ -956,12 +959,14 @@ class TestPositionPerformance(unittest.TestCase):
         )
 
         txn = create_txn(trades[1], 10.0, 1000)
-        pp = perf.PerformancePeriod(1000.0)
+        pt = perf.PositionTracker()
+        pp = perf.PerformancePeriod(1000.0, position_tracker=pt)
 
-        pp.execute_transaction(txn)
+        pt.execute_transaction(txn)
+        pp.handle_execution(txn)
 
         for trade in trades[:-2]:
-            pp.update_last_sale(trade)
+            pt.update_last_sale(trade)
 
         pp.calculate_performance()
 
@@ -989,7 +994,7 @@ class TestPositionPerformance(unittest.TestCase):
                       net_liquidation=1000.0)
 
         # now simulate a price jump to $11
-        pp.update_last_sale(trades[-1])
+        pt.update_last_sale(trades[-1])
 
         pp.calculate_performance()
 
@@ -1032,9 +1037,11 @@ class TestPositionPerformance(unittest.TestCase):
         )
 
         txn = create_txn(trades[1], 10.0, 100)
-        pp = perf.PerformancePeriod(1000.0)
+        pt = perf.PositionTracker()
+        pp = perf.PerformancePeriod(1000.0, position_tracker=pt)
 
-        pp.execute_transaction(txn)
+        pt.execute_transaction(txn)
+        pp.handle_execution(txn)
 
         # This verifies that the last sale price is being correctly
         # set in the positions. If this is not the case then returns can
@@ -1044,7 +1051,7 @@ class TestPositionPerformance(unittest.TestCase):
         self.assertEqual(pp.positions[1].last_sale_price, 10.0)
 
         for trade in trades:
-            pp.update_last_sale(trade)
+            pt.update_last_sale(trade)
 
         pp.calculate_performance()
 
@@ -1134,11 +1141,13 @@ single short-sale transaction"""
         trades_1 = trades[:-2]
 
         txn = create_txn(trades[1], 10.0, -100)
-        pp = perf.PerformancePeriod(1000.0)
+        pt = perf.PositionTracker()
+        pp = perf.PerformancePeriod(1000.0, position_tracker=pt)
 
-        pp.execute_transaction(txn)
+        pt.execute_transaction(txn)
+        pp.handle_execution(txn)
         for trade in trades_1:
-            pp.update_last_sale(trade)
+            pt.update_last_sale(trade)
 
         pp.calculate_performance()
 
@@ -1195,7 +1204,7 @@ single short-sale transaction"""
         pp.rollover()
 
         for trade in trades_2:
-            pp.update_last_sale(trade)
+            pt.update_last_sale(trade)
 
         pp.calculate_performance()
 
@@ -1249,15 +1258,17 @@ single short-sale transaction"""
         )
 
         # now run a performance period encompassing the entire trade sample.
-        ppTotal = perf.PerformancePeriod(1000.0)
+        ptTotal = perf.PositionTracker()
+        ppTotal = perf.PerformancePeriod(1000.0, position_tracker=ptTotal)
 
         for trade in trades_1:
-            ppTotal.update_last_sale(trade)
+            ptTotal.update_last_sale(trade)
 
-        ppTotal.execute_transaction(txn)
+        ptTotal.execute_transaction(txn)
+        ppTotal.handle_execution(txn)
 
         for trade in trades_2:
-            ppTotal.update_last_sale(trade)
+            ptTotal.update_last_sale(trade)
 
         ppTotal.calculate_performance()
 
@@ -1351,13 +1362,16 @@ trade after cover"""
         )
 
         cover_txn = create_txn(trades[6], 7.0, 100)
-        pp = perf.PerformancePeriod(1000.0)
+        pt = perf.PositionTracker()
+        pp = perf.PerformancePeriod(1000.0, position_tracker=pt)
 
-        pp.execute_transaction(short_txn)
-        pp.execute_transaction(cover_txn)
+        pt.execute_transaction(short_txn)
+        pp.handle_execution(short_txn)
+        pt.execute_transaction(cover_txn)
+        pp.handle_execution(cover_txn)
 
         for trade in trades:
-            pp.update_last_sale(trade)
+            pt.update_last_sale(trade)
 
         pp.calculate_performance()
 
@@ -1446,16 +1460,18 @@ shares in position"
         trades = factory.create_trade_history(*history_args)
         transactions = factory.create_txn_history(*history_args)
 
-        pp = perf.PerformancePeriod(1000.0)
+        pt = perf.PositionTracker()
+        pp = perf.PerformancePeriod(1000.0, position_tracker=pt)
 
         average_cost = 0
         for i, txn in enumerate(transactions):
-            pp.execute_transaction(txn)
+            pt.execute_transaction(txn)
+            pp.handle_execution(txn)
             average_cost = (average_cost * i + txn.price) / (i + 1)
             self.assertEqual(pp.positions[1].cost_basis, average_cost)
 
         for trade in trades:
-            pp.update_last_sale(trade)
+            pt.update_last_sale(trade)
 
         pp.calculate_performance()
 
@@ -1490,8 +1506,9 @@ shares in position"
 
         pp.rollover()
 
-        pp.execute_transaction(sale_txn)
-        pp.update_last_sale(down_tick)
+        pt.execute_transaction(sale_txn)
+        pp.handle_execution(sale_txn)
+        pt.update_last_sale(down_tick)
 
         pp.calculate_performance()
         self.assertEqual(
@@ -1509,19 +1526,22 @@ shares in position"
 
         self.assertEqual(pp.pnl, -800, "this period goes from +400 to -400")
 
-        pp3 = perf.PerformancePeriod(1000.0)
+        pt3 = perf.PositionTracker()
+        pp3 = perf.PerformancePeriod(1000.0, position_tracker=pt3)
 
         average_cost = 0
         for i, txn in enumerate(transactions):
-            pp3.execute_transaction(txn)
+            pt3.execute_transaction(txn)
+            pp3.handle_execution(txn)
             average_cost = (average_cost * i + txn.price) / (i + 1)
             self.assertEqual(pp3.positions[1].cost_basis, average_cost)
 
-        pp3.execute_transaction(sale_txn)
+        pt3.execute_transaction(sale_txn)
+        pp3.handle_execution(sale_txn)
 
         trades.append(down_tick)
         for trade in trades:
-            pp3.update_last_sale(trade)
+            pt3.update_last_sale(trade)
 
         pp3.calculate_performance()
         self.assertEqual(
@@ -1555,14 +1575,16 @@ shares in position"
         trades = factory.create_trade_history(*history_args)
         transactions = factory.create_txn_history(*history_args)
 
-        pp = perf.PerformancePeriod(1000.0)
+        pt = perf.PositionTracker()
+        pp = perf.PerformancePeriod(1000.0, position_tracker=pt)
 
         for txn, cb in zip(transactions, cost_bases):
-            pp.execute_transaction(txn)
+            pt.execute_transaction(txn)
+            pp.handle_execution(txn)
             self.assertEqual(pp.positions[1].cost_basis, cb)
 
         for trade in trades:
-            pp.update_last_sale(trade)
+            pt.update_last_sale(trade)
 
         pp.calculate_performance()
 
@@ -1713,12 +1735,16 @@ class TestPerformanceTracker(unittest.TestCase):
         self.assertEqual(perf_tracker.txn_count, len(txns))
         self.assertEqual(perf_tracker.txn_count, len(orders))
 
-        cumulative_pos = perf_tracker.cumulative_performance.positions[sid]
-        expected_size = len(txns) / 2 * -25
-        self.assertEqual(cumulative_pos.amount, expected_size)
+        positions = perf_tracker.cumulative_performance.positions
+        if len(txns) == 0:
+            self.assertNotIn(sid, positions)
+        else:
+            expected_size = len(txns) / 2 * -25
+            cumulative_pos = positions[sid]
+            self.assertEqual(cumulative_pos.amount, expected_size)
 
-        self.assertEqual(len(perf_messages),
-                         sim_params.days_in_period)
+            self.assertEqual(len(perf_messages),
+                             sim_params.days_in_period)
 
     def trades_with_txns(self, events, no_txn_dt):
         for event in events:
@@ -1844,3 +1870,32 @@ class TestPerformanceTracker(unittest.TestCase):
             # In the second bar we can start establishing a sharpe ratio.
             self.assertIsNone(msg_1['cumulative_risk_metrics']['sharpe'])
             self.assertIsNotNone(msg_2['cumulative_risk_metrics']['sharpe'])
+
+
+class TestPositionTracker(unittest.TestCase):
+    def setUp(self):
+        pass
+
+    def test_empty_positions(self):
+        """
+        make sure all the empty position stats return a numeric 0
+
+        Originally this bug was due to np.dot([], []) returning
+        np.bool_(False)
+        """
+        pt = perf.PositionTracker()
+
+        stats = [
+            'calculate_positions_value',
+            '_net_exposure',
+            '_gross_exposure',
+            '_short_exposure',
+            '_shorts_count',
+            '_long_exposure',
+            '_longs_count',
+        ]
+        for name in stats:
+            meth = getattr(pt, name)
+            val = meth()
+            self.assertEquals(val, 0)
+            self.assertNotIsInstance(val, (bool, np.bool_))
